db work

+++ db/query_interface.go
@@ -6,8 +6,8 @@ import (
 
 type QueriesInterface interface {
 	CreateUser(ctx context.Context, arg CreateUserParams) (User, error)
-	DeleteUser(ctx context.Context, id int32) error
-	GetUser(ctx context.Context, id int32) (User, error)
+	DeleteUser(ctx context.Context, id int64) error
+	GetUser(ctx context.Context, id int64) (User, error)
 	ListFiles(ctx context.Context) ([]File, error)
 	ListUsers(ctx context.Context) ([]User, error)
 	UpdatUser(ctx context.Context, arg UpdatUserParams) (User, error)
